Removed release-drafter + minor improvements
